Moved string manipulation code to an extension. Fixed a warning with string format

+++ Snap/Source/Game/Models/Cards/Suited Card/SuitedCard.m
@@ -77,7 +77,7 @@
             character = @"K";
             break;
         default:
-            character = [NSString stringWithFormat:@"%ld", rawValue];
+            character = [NSString stringWithFormat:@"%d", (int)rawValue];
             break;
     }
